optimized for EC2

+++ common-events/src/main/java/com/dops/common/event/InventoryReserved.java
@@ -1,9 +1,9 @@
 package com.dops.common.event;
 
 import lombok.Builder;
-
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Builder
-public record InventoryReserved(UUID orderId, UUID productId, Integer quantity, Double amount) {
+public record InventoryReserved(UUID orderId, UUID productId, Integer quantity, BigDecimal amount) {
 }
