+++ common-events/src/main/java/com/dops/common/event/OrderCreated.java
@@ -1,9 +1,9 @@
 package com.dops.common.event;
 
 import lombok.Builder;
-
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Builder
-public record OrderCreated(UUID orderId, UUID productId, Integer quantity, Double amount) {
+public record OrderCreated(UUID orderId, UUID productId, Integer quantity, BigDecimal amount) {
 }
