+++ inventory-service/src/main/java/com/dops/inventoryservice/dto/ProductResponse.java
@@ -2,7 +2,8 @@ package com.dops.inventoryservice.dto;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Builder
-public record ProductResponse(UUID productId, String sku, String productName, String description, Double price, Integer availableQuantity, String imageUrl) {}
+public record ProductResponse(UUID productId, String sku, String productName, String description, BigDecimal price, Integer availableQuantity, String imageUrl) {}
